refactor(admin): extract the Available Models panel from ConfigTab
First of the three independently-saving panels to leave ConfigTab.tsx
(plan 013, phase 1). Moves the Available Models panel and everything that
belongs only to it — handleSaveModel / handleDeleteModel / handleTestModel /
handleProbeContextWindow, the add-a-model wizard state, ModelTestDiagnostics,
DiagFact, ModelDraft, MODEL_PROVIDERS and inferProviderId — into
components/admin/config/ModelEditor.tsx.

Structural move only; no behavior change:

  - the panel keeps id="cfg-models" (the setup-checklist jump target)
  - modelTestResults / expandedModelTest stay keyed by model id, and the
    save / delete / test routes stay id-addressed (plan 011)
  - the tab-level error banner is still shared: `error` in, `onError` out, so
    a model error still surfaces at the top of the tab and the wizard still
    renders it inline
  - the first-run "connect your first model" jump keeps its exact guard
    (skip when a model exists or the form is already open) via a small
    imperative handle, since only the panel knows the form is open

Six props, no setters passed down. The shared inline-style objects moved to
config/styles.ts verbatim rather than being threaded through six style props —
this tree stays inline-styled by design.

ConfigTab.tsx: 2,875 → 1,957 lines. The sticky-save cluster and
handleSaveConfig are untouched (phase 2). All 11 ConfigTab characterization
tests still pass unchanged, plus 11 new tests exercising the panel in
isolation. 47 files / 357 tests; ConfigTab chunk 87,775 → 87,780 bytes.
